presentation folder cleanup
