--- src/main/java/presentation/controller/MainController.java
@@ -1,4 +1,0 @@
-package presentation.controller;
-
-public class MainController {
-}
