+++ src/main/java/presentation/controller/SignupController.java
@@ -4,8 +4,8 @@ import domain.model.Role;
 import domain.model.RoleType;
 import domain.model.User;
 import domain.service.PasswordService;
-import infrastructure.repository.UserRepository;
 import infrastructure.repository.RoleRepository;
+import infrastructure.repository.UserRepository;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import presentation.view.SceneManager;
