+++ src/main/java/presentation/view/SceneManager.java
@@ -1,25 +1,30 @@
 package presentation.view;
-import domain.model.StudyMaterial;
 
+import domain.model.Category;
+import domain.model.StudyMaterial;
 import domain.model.Tag;
 import domain.model.User;
-import domain.model.Category;
-
-import domain.service.Session;
+import infrastructure.repository.CategoryRepository;
 import infrastructure.repository.StudyMaterialRepository;
 import infrastructure.repository.TagRepository;
 import infrastructure.repository.UserRepository;
-import infrastructure.repository.CategoryRepository;
-
-import javafx.scene.Node;
-import javafx.scene.control.*;
-import javafx.scene.layout.*;
-
 import javafx.fxml.FXMLLoader;
 import javafx.geometry.Insets;
+import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.Hyperlink;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
+import presentation.components.CategoryPage;
+import presentation.components.ListItem;
+import presentation.controller.*;
+import presentation.utility.GUILogger;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -27,9 +32,6 @@ import java.util.List;
 import java.util.Objects;
 
 import static presentation.view.Screen.*;
-
-import presentation.components.*;
-import presentation.utility.GUILogger;
 
 public class SceneManager {
     private static SceneManager instance;
@@ -53,16 +55,15 @@ public class SceneManager {
     }
 
     private void initializeComponents() throws IOException {
-        instance.current = FXMLLoader.load(Objects.requireNonNull(SceneManager.class.getResource("/fxml/login.fxml")));
+        setLogin();
         instance.header = FXMLLoader.load(Objects.requireNonNull(SceneManager.class.getResource("/fxml/header.fxml")));
         instance.footer = FXMLLoader.load(Objects.requireNonNull(SceneManager.class.getResource("/fxml/footer.fxml")));
         instance.logged = false;
     }
 
     public void displayCategory(int id) {
-        if (Session.getInstance().getCurrentUser() == null){
-            try { setScreen(SCREEN_LOGIN); }
-            catch (IOException e){ throw new RuntimeException("Failed to load FXML components", e);}
+        if (CurrentUserManager.get() == null){
+            setLogin();
             return;
         }
 
@@ -78,9 +79,8 @@ public class SceneManager {
     }
 
     public void displayMaterial(int id) {
-        if (Session.getInstance().getCurrentUser() == null){
-            try { setScreen(SCREEN_LOGIN); }
-            catch (IOException e){ throw new RuntimeException("Failed to load FXML components", e);}
+        if (CurrentUserManager.get() == null){
+            setLogin();
             return;
         }
 
@@ -93,14 +93,13 @@ public class SceneManager {
             return;
         }
 
-        MaterialPage page = new MaterialPage(s);
+        StudyMaterialPageController page = new StudyMaterialPageController(s);
         page.displayPage();
     }
 
     public void displayProfile(int id) {
-        if (Session.getInstance().getCurrentUser() == null){
-            try { setScreen(SCREEN_LOGIN); }
-            catch (IOException e){ throw new RuntimeException("Failed to load FXML components", e);}
+        if (CurrentUserManager.get() == null){
+            setLogin();
             return;
         }
 
@@ -113,7 +112,7 @@ public class SceneManager {
             return;
         }
 
-        ProfilePage.setPage(u);
+        ProfilePageController.setPage(u);
     }
 
     public void displayErrorPage(String errorText, Screen redirectScreen, String redirectLabel) {
@@ -167,9 +166,14 @@ public class SceneManager {
         instance.current.setCenter(vbox);
     }
 
+    public void setLogin(){
+        LoginController lPage = new LoginController();
+        lPage.initialize(instance.current);
+    }
+
     public void setScreen(Screen screen) throws IOException {
         if (!instance.logged){
-            instance.current = FXMLLoader.load(Objects.requireNonNull(SceneManager.class.getResource(screen == SCREEN_SIGNUP ? "/fxml/signup.fxml" : "/fxml/login.fxml")));
+            setLogin();
         } else {
             BorderPane bp = new BorderPane();
             bp.setTop(instance.header);
@@ -188,11 +192,11 @@ public class SceneManager {
                     base.setContent(FXMLLoader.load(Objects.requireNonNull(SceneManager.class.getResource("/fxml/search.fxml"))));
                     break;
                 case SCREEN_PROFILE:
-                    MyProfilePage pPage = new MyProfilePage();
+                    MyProfileController pPage = new MyProfileController();
                     pPage.initialize(base);
                     break;
                 case SCREEN_UPLOAD:
-                    UploadPage uPage = new UploadPage();
+                    UploadController uPage = new UploadController();
                     uPage.initialize(base);
                     break;
                 default:
@@ -231,7 +235,7 @@ public class SceneManager {
     public void logout() throws IOException {
         if (instance.logged){
             instance.logged = false;
-            instance.setScreen(SCREEN_LOGIN);
+            setLogin();
         } else {
             GUILogger.warn("User is already logged out");
         }
